+++ ffhs-moto-app/src/components/Register.js
@@ -6,53 +6,44 @@ import '../css/style.css';
 
 // import
 import RegisterForm from './RegisterForm';
+import LoginForm from './LoginForm';
 
 // create hook
 const RegisterBody = () => {
 
-  // dummy-user ****delete afterwards****
-  const adminUser = {
-    name: "admin",
-    email: "admin@admin.com",
-    password: "admin123"
-  }
-
-  // create states
-  const [user, setUser] = useState({name: "", email: ""});
+  // create states (email not necessary probably)
+  const [user, setUser] = useState({userId: "", name: "", email: ""});
   const [error, setError] = useState("");
 
-  // 
-  const Login = (details) => {
-    console.log(details);
-    setUser({
-      name: details.name,
-      email: details.email
+  // create Register request, setStates with received data
+  const Register = (data) => {
+    fetch('http://localhost:3000/api/user/register', {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(data),
     })
-
-    if(details.email == adminUser.email && details.password == adminUser.password) {
-      console.log("Logged in");
-    } else {
-      setError("Details do not match")
+    .then(response => response.json())
+    .then(json => {
+      setError(json.error)
+      if(json.error === null) {setUser({email:json.data.email, userId: json.data.userId, name: json.data.name}) }
+      console.log(json)
+    })
+    .catch(err => alert(err));
     }
-  }
-
-
-  //
-  const Logout = () => {
-    setUser({ name: "", email: ""});
-  }
 
   // if user.email state is not empty show text and logout, else show login form
   return (
     <React.Fragment>
       <div className="App">
-        {(user.email != "") ? (
-          <div className = "welcome">
-            <h2>Welcome, <span>{user.name}</span></h2>
-            <button onClick={Logout}>Logout</button>
+        {(user.userId != "") ? (
+          <div className = "registersuccess">
+            <h2>You have been successfully registered, <span>{user.name}</span></h2>
+            <LoginForm />
           </div>
         ) : (
-          <RegisterForm Login={Login} error={error} />
+          <RegisterForm Register={Register} error={error} />
         )}
       </div>
     </React.Fragment>
